requirements.txt
streamlit==1.34.0
boto3==1.34.59
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
mplfinance==0.12.10b0
yfinance==0.2.38
PyPortfolioOpt==1.5.5
python-dotenv==1.0.1